+++ index.ts
@@ -1,1 +1,3 @@
-export * from './src/BrazeMessageComponent';
+export * from './banner-index';
+export * from './end-of-article-index';
+export * from './logic-index';
