+++ package.json
@@ -10,6 +10,8 @@
   "types": "dist/index.d.ts",
   "files": [
     "dist",
+    "banner/package.json",
+    "end-of-article/package.json",
     "logic/package.json"
   ],
   "scripts": {
